Remove feature, move test to own file
